requirements.txt
flask
flask-cors
pandas
google-generativeai
sentence-transformers
scikit-learn
python-dotenv